+++ shaders/pfshader.glsl
@@ -9,4 +9,8 @@ void main() {
 
     fragcolor = texture(texSampler, texCoord);
 
+    if (fragcolor.a <= 0.2) {
+        discard;
+    }
+
 }